SingleGAN-master/requirements.txt
absl-py==0.9.0
appdirs==1.4.3
asn1crypto==1.2.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
calmsize==0.1.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.13
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
dominate==2.4.0
easydict==1.9
entrypoints==0.3
filelock==3.0.12
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
menuinst==1.4.16
mistune==0.8.4
more-itertools==8.0.2
msgpack==0.6.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.3
nibabel==2.5.0
notebook==6.0.1
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.2
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-memlab==0.0.4
pytz==2019.2
PyWavelets==1.0.3
pywin32==225
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-image==0.16.2
scipy==1.3.3
Send2Trash==1.5.0
SimpleITK==1.2.4
six==1.13.0
tensorboard==2.0.0
tensorboardX==1.9
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchfile==0.1.0
torchvision==0.4.2
torchviz==0.0.1
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
urllib3==1.24.2
virtualenv==20.0.10
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
zipp==0.6.0
